Backend/requirements.txt
wavio
scipy
soundfile
g2p-en
torch
torchvision
torchaudio
sounddevice
accelerate
pyaudio
google-generativeai
fastapi
uvicorn
python-dotenv
transformers
pymilvus
sentence-transformers